Significantly restructure pie-3d add() validation
Now it's starting to take shape: the validation module provides a
function to get the validation dictionary (the purpose is to have an
ability to provide some dependencies to the dictionary, e.g. the
HTMLDivElement class, which won't be available in the unit tests, so
must be injected). Also these changes required changes in types, which
in turn required changes in other modules. Currently the tests (except
the skipped ones) pass.

Also now the pie-3d validation module doesn't expect the propsToCheckSet
to be passed to it, but instead determines those on its own, by taking
the dictionary from the ./src/validation/add-method-arg.js module and
excluding all the props present in it from the user provided argument's
props.