Use new approach in .add() arg common validation
The new approach is to have a validation module, which provides a
validation function and a function which returns ValidationDictionary.
This dictionary is then passed to the validation function. In the
modified module it's not necessary, but in other places it is. And for
uniformity it's been decided to use it here too.