Do Flow build
This finishes rework to the new validation approach:
  - have validation modules, which provide a dictionary getter and a
    validate function
  - create the dictionary (inject dependencies to it if any) and pass it
    to the validate() function
  - in the validate function do preliminary checks if any, then use the
    ./src/validation/by-dictionary.js module, then post checks if any.

This allows to use the ./src/validation/by-dictionary.js module in
virtually any place in the library. And this is supposed to be
convenient, since the module checks for the presence, absence of
properties, their type/value, missing properties etc.

Now the validation for the new XCharts() call has to be done, which is
also supposed to utilize that by-dictionary module.