Don't store constructor arg, store container div
It's been decided to to save a reference to the constructor argument. No
point to do it, b/c since we don't freeze it there is no guarantee that
it won't be changed outside of the xCharts instance, also it potentially
can prevent garbage collection. The container div is stored instead,
which can make sense, b/c we need it and thru the xCharts instance's it
anyway should stay.

Also a Flow build has been done.