Add/tweak skipped tests
The issue has been reproduced with a different canvas dimensions, chart
coordinates, to facilitate debugging (it's more clear what's wrong when
the chart is centerX rotated further to be more 'flat'). Originally
those skipped tests had the pie chart rotated to 153 deg. which made it
more close to a circle than an ellipse. It was not convenient.