Fix test util
The util in question didn't draw slices on a rim correctly if the tails
was visible. B/c we draw this way: from the point at the left edge on
visible face, to the left edge point on invisible face, then the ellipse
on the invisible face (clockwise if the rim is visible at the top,
counterclockwise if at the bottom), then end point on the visible face
and ellipse back to the point where we started.

The rim is visible at the top only if the tails is visible, b/c for
angles > 180⁰ we fully reverse the pie and the heads becomes tails etc.
(Take the lid from a can and manipulate it to get an idea).