Add & amend skipped test
In order to facilitate debugging the test argument has been changed (to
be more convenient to work in the test utils).

Also select the name for the test, b/c the fix is already done (not
committed yet), and we know what's the reason for the wrong behavior and
what's relevant in the test argument.