Hard code versions of flow npm modules
Currently I use a Linux system which doesn't have a required glibc
version, so I can't install the latest versions of those modules. I used
to manually monitor that those are not installed, but now it occurred to
me that it's better to temporarily hard code the versions.

The Fedora image actually seemed to allow the installation and
everything could work, probably. But in my text editor, which uses those
modules, but installed on the host machine (for various stuff like an
LSP server etc.) only the old versions are available. So for consistency
I want the versions in the image (which are used for builds from the
flow files etc.) and on my host machine (used for LSP etc.) to match.

The image is supposed to be rebuild on this commit, b/c the package.json
file has been changed. In reality it was not and the image tag was
changed afterwards and subsequent commits rebased.