Remove Makefile targets for copy to/from container
It's easier to run 'make podman-container-bash' and manipulate the files
from the container then to do that mambo-jumbo with variables etc. I
haven't used this target in a while and it assumes the 1-to-1
correspondence in locations between the local file system and in the
container. If you need to copy a file from ./local/foo into the
/container/foo/bar location — this target is useless.