Add isComponentInspectMode option
The general idea is to implement the legend. It's been decided that it
will be done this way:

  - the user will provide HTML/CSS which will be inserted to its
    shadowDom as is
  - the browser will render it
  - the user will be responsible for the valid markup, CSS properties
    etc.

In order to facilitate the debuggin of this process, it's been decided
to remove the 'cover canvas' (which is present in order to allow one day
to save the whole chart with all components as a single image).

So this options was introduced. Currently it's only allowed, not handled
in any way except its validation.