Rename types and methods, do Flow build
The issue was this: we have an instance of the XCharts class, which has
the .add() method. Yet the type and various related methods were named
as addComponentArgument-something and similar. Didn't look good.

So I've decided to rename the types and methods, instead of renaming the
.add() method to an .addComponent() method.

Also in some cases types were imported but not used. Removed them w/o
creating a dedicated commit.