Rework demo example JS
To make it more presentable to the user, so that the user could see
relevant configuration options at the top of the JS string, since he'll
be viewing it in that LiveJsEditor, which is not an IDE and doesn't
allow to quickly jump to function/method definitions etc.

Also removed the mention of the width/height to be set on the container
div. Everything seems to work w/o it. Will consider dealing with this
later: either throw an error in this case, or explicitly document it or
just ignore.