Make LiveJsEditor height fixed and resizable
It turned out that if there is a lot of code in the pre elements, then
the height of the pre block is so big, that it's inconvenient to the
user to make changes then scroll top to press the Run button.

So the intention was to allow the user make a change, then w/o any
scroll press the Run button and see the resulting chart.