Remove almost everything from footer in docs
There are not that many things to show in the footer, specifically
couldn't find a proper name for the Docs section in the footer. I's
about to have a page titled Docs, which would contain everything:
general notes + API. So how would it look in the footer:

Docs
  Docs

So I decided just to remove almost everything from the footer, except
for the copyright note.