Attempt to fix demo page
NOTE: the docusaurus build has not been done yet, will be done after
other issues are solved (the demo isn't working in Firefox).

The page is: https://koroliov.github.io/x-charts-js/docs/demos/
The problem is that the library js modules are fetched with an incorrect
URL. Probably this is due to the relative paths, b/c:

- at a certain moment, when the user enters the
  https://koroliov.github.io/x-charts-js/ he is redirected to
  https://koroliov.github.io/x-charts-js/docs/demos/

at this point locally the relative path is resolved correctly, yet on
GitHub — it's not, probably b/c of some timing issues. Now it should
work, although I have to push it to GitHub to see if it's true.