Fix errors in FF
In FF the previous code didn't work. With the NoScript plugin (page
trusted, of course) it didn't work at all. With no plugin it didn't auto
run the code. Only after the Run button press, if you visit the demo
page directly, not the root page and redirect to the demo, it worked.

Now it works in FF, Edge, Chrome.