Remove dev server's SSL stuff
After the repository went public I received an alert from
GitGuardian.com about that stuff. A decision has been made to remove it
in HEAD, gitignore. The alert to be marked as 'resolved' with a reason
like 'testing certificate'.

Then openssl to be added to the container and a make target added to
generate a new dev cert.

This commit just removes the already committed files.