Reorganize various in-container stuff
The image hasn't been rebuilt since this is an intermediary commit.

The intention is to generate the server certificates on image build and
store them in the container. For this purpose it's been decided not to
introduce numerous directories in /home/x-charts-js directory, but store
all of them under a single directory: var-contained (i.e. not mounted,
not shared with the host machine).

So now this has been done and this directory's name has been defined as
a variable and referenced in the package.json file too.