Create certificate on image build, use microdnf
Now the server should start (in several previous commits that wasn't the
case). The certificate and private key will be generated on image build
and stored in the container, which should resolve the GitGuardian
message.

The microdnf change has also been committed, although it was present
locally for a while now (like the previous 3 commits, approx.).