Remove lots of sandbox stuff in docs
That stuff was generated by Docusarus as its example/tutorial website.
Originally I left it, since it seemed more handy to have it, e.g. to see
which formatting options are available etc. But it turned out that it's
still present in the sitemap.xml and could be accessed by a direct link,
indexed etc.