Add script to enforce version matches
Since we have several places where the version number is present:

- ./flow/src/version.js
- ./dist/modules/version.js
- ./docs-src/static/x-charts-js/version.js
- git tag
- changelog.md

And not mentioning the released zip file and the build ./docs, it's been
decided to somehow check for the match.

To check git tag was considered to be an overkill, since it would
require local git hooks etc, I see no need for it currently.

But on a Docusaurus build there is a check now, that values should match
in:

- changelog.md,
- ./dist/modules/version.js and
- ./docs-src/static/x-charts-js/version.js

The idea is that when a tag with the version is created, the developer
absolutely will check at least 1 of those places (very likely the built
docs, i.e. the value taken from the changelog.md). And if the developer
sees that in the built docs the value is correct, then the whole 3
places are correct.

Looks more or less reliable to me.