Rename meta to options in pie-3d addMethodArg
The intention is that one day we may add some way to override options on
a particular data element, e.g. radius of the circle etc.

And I just didn't like that 'meta'.

The Flow build has been done.

Also introduced a file which lists changes applied in the code, but not
yet documented.