On pie-3d make stroke color & width configurable
Also there was an idea to implement default options, since it wouldn't
look good to require the user to provide strokeColor if he wants
strokeWidth = 0. But this was decided to be postponed.

It's not that easy to rework the validation mechanism and allow it to
provide default values too. And since:

- it wouldn't give a significantly noticeable benefit to the library (in
  comparison to animations, interactivity)

- plus in demos there will be a need to use all options, to allow a
  potential adopter see which configuration options are present

it made sense to deprioritize this

So a situation, where the user will be required to provide options like:

  rotationAroundCenterXAxis: 30,
  strokeColor: '#000000',
  strokeWidth: 0,

was deemed a lesser evil.