Update npm modules, increment image tags
The docs image might not need an increment, but I've decided to do it,
since I've built it many times today and not sure if it contains any
changes.

An attempt to upgrade Flow to the latest version 0.291.0 failed, the DOM
types and probably other were missing, so had to downgrade back to
0.259.1. Meanwhile though, reinstalled npm modules in the main image, so
the package-lock.json file is changed.

Everything seems to work, with the latest versions, except Flow. Created
a bug for that: https://github.com/facebook/flow/issues/9335