Upgrade Flow v0.259.1 -> v0.291.0, rebuild images
Currently there are numerous type errors, but those are related only to
this repo's source code, i.e. no missing, wrong types outside of this
project's control.

B/c of this the unit tests & Flow build don't pass, but that's fine. In
subsequent commits the type errors will be dealt with.

Installed the flow-typed npm module, b/c since approximately version
0.261.0 Flow stopped to ship various 'global' types: DOM, HTML, Nodejs
etc, which now are supposed to be added via that module.