Change SELinux setup
A problem was encountered that if a file in a bind mounted volume was
'git restore'-d then the mounted directory became inaccessible from the
container, which broke a lot of Makefile targets etc.

The solution chosen is to disable labeling in the compose file. This
seems to be acceptable, since we don't have any secrets, the containers
are used only for local development and Podman is rootless.