Reorganize file structure
There are modules, e.g. ./flow/src/validation/add-method-arg.js, which
are related only to the main exported class.

Other modules, e.g.

  ./flow/src/validation/by-dictionary.js
  ./flow/src/types.js
  ./flow/src/utils/

are different, b/c they are used globally in many other modules.

The intention of this commit it to start fixing this. The validation
modules have been left in place (since they will be gone by the end of
this branch, when the external-values-processor module/modules are
implemented).