Rename test-unit/utils to test-unit/test-utils
I didn't like the fact, that we have src/utils and in the
test-unit/utils there were no tests at all, although one might expect
that it contained tests for the src files.

This seemed to be the most clean way to deal with the situation: just
have a location in the test-unit directory, which would 100% never have
a corresponding entry in the src directory.