Rework isObject() util
This util should return true, if the value passed is either declared as
an object literal or an object created with Object.create(null);

This seems to be not possible to detect in every possible case, e.g. if
something was instantiated as an instance of a class, then its proto set
to null etc. But these are edge cases unlikely to happen.

The overall goal is to facilitate validation of some external values,
where the passing party is likely to make mistakes and pass something
different where only a {  } or Object.create(null) are expected. For
this goal the added changes should be good and useful.