Add trailingSlash=true option
The purpose of this change is an attempt to solve a problem with the
https://koroliov.github.io/x-charts-js/sitemap.xml submitted to Google
Search Console. After multiple submits its 'type' column value is
Unknown.

May be the fact that pages served have urls (in the browser's address
bar) like:

  https://koroliov.github.io/x-charts-js/docs/api-ref/components/legend/

yet in the sitempa.xml they were presented as:

  https://koroliov.github.io/x-charts-js/docs/api-ref/components/legend

If this change will solve the problem  is not known, this is an attempt,
but w/o a commit/push to GitHub there is now way to make the sitemap.xml
file to change on server.

Also the Docusaurus build will be performed separately.