Rework index and add overview pages
The purpose of this change is to address problems with Google search.
Currently no documentation pages are indexed. The sitemap.xml is not
accepted (the error is "Couldn't fetch"), which is probably due to the
non indexing issue. The Google Search Central docs mention that such a
couldn't fetch error can happen if Google sees no need in the sitemap,
as I understood, if it doesn't consider the site/links relevant.

So this is an attempt to fix that.

An idea to restore the default Docusaurus index page was dismissed,
since as of now the library has very little content to put to such an
index page (no features etc.). So the current index page with a
JavaScript redirect has remained, but the redirect now points to the
/x-charts-js/docs/overview/ page, which has been created by combining
all the content from the /x-charts-js/docs/overview/ directory into a
single page, with mostly looks&feels changes.