Add global type PureObject
The idea is this: we need somehow to refer to objects which exist purely
as 'data'. I.e. it's not an instance of some class which handles
behavior etc. It only provides either data or 'configuration options'
etc. The 'configuration options' can be functions, e.g. the user passes
an object with a function to a validator and the validator's
behavior depends on that function. May be they can be even instances of
some classes (although I don't have such a situation now), but the root
object, the tested value, is not an instance, it's a pure dictionary of
some values.

Examples:

  const pureObj = {}

  const pureObj = {
    foo() {},
    bar: '123',
  }

  const pureObj = Object.create(null);
  pureDict.foo = '123';

The need for this type exists, b/c it's going to be used widely
throughout the code in various checks.