Move component registered check to all validation
I.e. I've been working on the external-values-processor, which is
supposed to do validation too, along with the default values provision
etc. And basically, it's better to move the
ERR_X_CHARTS_JS_COMPONENT_NOT_REGISTERED error check there also.

Currently this is not validated, b/c only the happy path cases have been
done, but it will be.