Restore _attemptToShowError() call, do Flow build
This is how it worked and it was removed b/c I incorrectly thought that
it's never possible to show the error on the page, b/c the initDom()
call came later in the code. The fact that it comes later doesn't
prevent the error to be shown on the page in all cases. In some cases
it's still possible.