Determine props handled on main level dynamically
That is, when the user passes an argument, it's first
validated/processed on the main level (currently only for the type and
zIndex properties), since some properties will be used in all
components, then it's passed for processing/validation to the component.
The already checked properties should be ignored and not cause the
'unknown property' validation error. In order to determine them
dynamically changes had to be made to the main level processor, which
now exports its argument's schema.

Refactoring is still needed in the
  ./flow/src/external-values-processor/main.js.